repackage event handling to bind on document, not 'body'
error events now send an object, which contains the tiddler
and some context information. I chose not to have a different
error 'type' for each because that seemed noisy, therefore the
data passed requires the context. I'm not satisfied with this
so will probably change it yet again, but still experimenting.

More to do.

thanks to @fnd for feedback